refact: change file tree for delivery on server side
