+++ vsmtp-server/src/tests/mail_queue/deliver/test
@@ -1,1 +1,1 @@
-{"connection_timestamp":{"secs_since_epoch":1649169410,"nanos_since_epoch":301304167},"client_addr":"127.0.0.1:80","envelop":{"helo":"client.com","mail_from":"from@client.com","rcpt":[{"address":"to+1@client.com","transfer_method":"Deliver","email_status":"Waiting"},{"address":"to+2@client.com","transfer_method":"Maildir","email_status":"Waiting"}]},"body":{"Parsed":{"headers":[["date","bar"],["from","foo"]],"body":{"Regular":[]}}},"metadata":{"timestamp":{"secs_since_epoch":1649169410,"nanos_since_epoch":301539989},"message_id":"test","skipped":null}}
+{"connection_timestamp":{"secs_since_epoch":1649171639,"nanos_since_epoch":202175876},"client_addr":"127.0.0.1:80","envelop":{"helo":"client.com","mail_from":"from@client.com","rcpt":[{"address":"to+1@client.com","transfer_method":"Deliver","email_status":"Waiting"},{"address":"to+2@client.com","transfer_method":"Maildir","email_status":"Waiting"}]},"body":{"Parsed":{"headers":[["date","bar"],["from","foo"]],"body":{"Regular":[]}}},"metadata":{"timestamp":{"secs_since_epoch":1649171639,"nanos_since_epoch":202576677},"message_id":"test","skipped":null}}
